feat(shared): extract ASSESSMENT_BUCKETS to shared/domain/assessment.ts; delete dead barrel files
- Move ASSESSMENT_BUCKETS + AssessmentBucketKey from api/Services/Course/buckets/assessment.ts
  to shared/domain/assessment.ts; update api importers in CourseFacetService + CourseFilterBuilder
- Delete scraper/src/types/{insis,jobs,queue}.ts — all were single-line re-exports of shared/ modules
  with zero source importers
