--- scraper/src/types/insis.ts
@@ -1,3 +1,0 @@
-export * from '@shared/domain/insis'
-export * from '@shared/queue/insis'
-export type { InSISDay as default } from '@shared/domain/insis'
