--- scraper/src/types/jobs.ts
@@ -1,1 +1,0 @@
-export * from '@shared/queue/jobs'
