--- scraper/src/types/queue.ts
@@ -1,1 +1,0 @@
-export * from '@shared/queue/names'
